Revert "move to Dapper.AOT.Test.Integration.csproj"
This reverts commit e80464847eb93880d6b4c36227a32897281b32f8.
