--- test/Dapper.AOT.Test.Integration/Dapper.AOT.Test.Integration.csproj
@@ -1,54 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFrameworks>net6.0;net48</TargetFrameworks>
-        <RootNamespace>Dapper.AOT.Test.Integration</RootNamespace>
-        <DefineConstants>$(DefineConstants);DAPPERAOT_INTERNAL</DefineConstants>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <Compile Remove="Accessors/Data/**/*.*.cs" />
-        <None Include="Accessors/Data/**/*.*" />
-        <None Update="Accessors/Data/**/*.*" CopyToOutputDirectory="PreserveNewest" />
-
-        <Compile Remove="Interceptors/**/*.*.cs" />
-        <None Include="Interceptors/**/*.*" />
-        <None Update="Interceptors/**/*.*" CopyToOutputDirectory="PreserveNewest" />
-
-        <Compile Include="..\..\src\Dapper.AOT.Analyzers\AotGridReader.cs" Link="AotGridReader.cs" />
-        <Compile Include="..\..\src\Dapper.AOT.Analyzers\InGeneration\DapperHelpers.cs" Link="DapperHelpers.cs" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Dapper" Condition="'$(TargetFramework)'!='net48'" />
-        <PackageReference Include="Dapper.StrongName" Condition="'$(TargetFramework)'=='net48'" />
-        <PackageReference Include="System.ComponentModel.Annotations" Condition="'$(TargetFramework)'=='net48'" />
-        <PackageReference Include="Microsoft.Build" Condition="'$(TargetFramework)'!='net6.0'" />
-        <PackageReference Include="Microsoft.Build" VersionOverride="[17.3.2]" Condition="'$(TargetFramework)'=='net6.0'" />
-        <PackageReference Include="Microsoft.Build.Utilities.Core" />
-        <PackageReference Include="Npgsql" />
-        <PackageReference Include="Oracle.ManagedDataAccess" Condition="'$(TargetFramework)'=='net48'" />
-        <PackageReference Include="Oracle.ManagedDataAccess.Core" Condition="'$(TargetFramework)'!='net48'" />
-        <PackageReference Include="System.ComponentModel.Annotations" />
-        <PackageReference Include="System.Data.Common" />
-        <PackageReference Include="System.Data.SqlClient" />
-        <PackageReference Include="Microsoft.Data.SqlClient" />
-        <PackageReference Include="xunit" />
-        <PackageReference Include="xunit.runner.visualstudio">
-            <PrivateAssets>all</PrivateAssets>
-            <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
-        </PackageReference>
-        <PackageReference Include="Microsoft.NET.Test.Sdk" />
-        <ProjectReference Include="../../src/Dapper.AOT.Analyzers/Dapper.AOT.Analyzers.csproj" />
-        <ProjectReference Include="../../src/Dapper.AOT/Dapper.AOT.csproj" />
-
-        <PackageReference Include="Microsoft.CodeAnalysis.CSharp.Workspaces" />
-        <PackageReference Include="Microsoft.CodeAnalysis.VisualBasic.Workspaces" />
-        <PackageReference Include="Microsoft.CodeAnalysis.CSharp.Analyzer.Testing.XUnit" />
-        <PackageReference Include="Microsoft.CodeAnalysis.CSharp.CodeFix.Testing.XUnit" />
-        <PackageReference Include="Microsoft.CodeAnalysis.VisualBasic.Analyzer.Testing.XUnit" />
-        <PackageReference Include="Microsoft.CodeAnalysis.VisualBasic.CodeFix.Testing.XUnit" />
-        <PackageReference Include="Testcontainers.PostgreSql" />
-    </ItemGroup>
-    
-</Project>
