+++ test/Dapper.AOT.Test/Integration/PostgresqlFixture.cs
@@ -1,9 +1,9 @@
-﻿using System.Threading.Tasks;
-using Npgsql;
+﻿using Npgsql;
+using System.Threading.Tasks;
 using Testcontainers.PostgreSql;
 using Xunit;
 
-namespace Dapper.AOT.Test.Integration.Setup;
+namespace Dapper.AOT.Test.Integration;
 
 [CollectionDefinition(Collection)]
 public class SharedPostgresqlClient : ICollectionFixture<PostgresqlFixture>
