+++ test/Dapper.AOT.Test/Integration/QueryTests.cs
@@ -3,7 +3,6 @@ using System;
 using System.Collections.Generic;
 using System.Data.Common;
 using System.Threading.Tasks;
-using Dapper.AOT.Test.Integration.Setup;
 using Xunit;
 
 namespace Dapper.AOT.Test.Integration;
